# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Display Management API
  version: 2.0.0
  description: |
    API для управления домами, дисплеями, группами, шаблонами и экстренными сообщениями.

    ## Аутентификация
    Все эндпоинты (кроме `/auth/*`) требуют JWT Bearer токен в заголовке:
    ```
    Authorization: Bearer <accessToken>
    ```
servers:
  - url: https://api.deeply-bath.space
    description: Production
  - url: http://localhost:8082
    description: Development
security:
  - BearerAuth: []
...
paths:
...
  /complexes/{complexId}/houses/{id}/displays:
    parameters:
      - $ref: '#/components/parameters/ComplexId'
      - $ref: '#/components/parameters/HouseId'

    get:
      tags: [Displays]
      summary: Получить список дисплеев дома
      operationId: getDisplays
      responses:
        '200':
          description: Список дисплеев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Display'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Displays]
      summary: Создать дисплей
      operationId: createDisplay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisplayInput'
      responses:
        '201':
          description: Дисплей создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Display'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /complexes/{complexId}/houses/{id}/displays/{displayId}:
    parameters:
      - $ref: '#/components/parameters/ComplexId'
      - $ref: '#/components/parameters/HouseId'
      - $ref: '#/components/parameters/DisplayId'

    get:
      tags: [Displays]
      summary: Получить дисплей по ID
      operationId: getDisplayById
      responses:
        '200':
          description: Дисплей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Display'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Displays]
      summary: Обновить дисплей
      operationId: updateDisplay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisplayInput'
      responses:
        '200':
          description: Дисплей обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Display'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Displays]
      summary: Удалить дисплей
      operationId: deleteDisplay
      responses:
        '204':
          description: Дисплей удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ComplexId:
      name: complexId
      in: path
      required: true
      description: ID жилого комплекса
      schema:
        type: integer
        format: int64
    HouseId:
      name: id
      in: path
      required: true
      description: ID дома
      schema:
        type: integer
        format: int64
    DisplayId:
      name: displayId
      in: path
      required: true
      description: ID дисплея
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Токен отсутствует или недействителен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Display:
      type: object
      required: [id, name, house_id, token]
      properties:
        id:
          type: integer
          format: int64
          example: 10
        token:
          type: string
          example: "8S1mv0lv"
        name:
          type: string
          example: "Дисплей подъезд 1"
        template_id:
          type: integer
          format: int64
          example: 3
        house_id:
          type: integer
          format: int64
          example: 1
        floor:
          type: integer
          nullable: true
          example: 3
        entrance:
          type: integer
          nullable: true
          example: 1
    DisplayInput:
      type: object
      required: [name, house_id]
      properties:
        name:
          type: string
          example: "Дисплей подъезд 1"
        template_id:
          type: integer
          format: int64
          example: 3
        house_id:
          type: integer
          format: int64
          example: 1
        floor:
          type: integer
          nullable: true
          example: 3
        entrance:
          type: integer
          nullable: true
          example: 1
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Ресурс не найден"